system-monitor/requirements.txt
pyserial
rich
dotenv
keyboard